Move: roles enums to shared
